% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_artist_names.R
\name{extract_artist_names}
\alias{extract_artist_names}
\title{To extract artist names from text.}
\usage{
extract_artist_names(api_key, text, min_hotttnesss = NA,
  max_hotttnesss = NA, min_familiarity = NA, max_familiarity = NA,
  sort = NA, results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{text}{text that contains artist names}

\item{min_hotttnesss}{the minimum hotttnesss for returned artists}

\item{max_hotttnesss}{the maximum hotttnesss for returned artists}

\item{min_familiarity}{the minimum familiarity for returned artists}

\item{max_familiarity}{the maximum familiarity for returned artists}

\item{sort}{specified the sort order of the results}

\item{results}{the number of results desired}
}
\value{
data frame giving artist's names
}
\description{
To extract artist names from text.
}
\examples{
\dontrun{
data=extract_artist_names(api_key,text="I like adele and Maroon 5")
}
}

