% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest.R
\name{get_artist_blogs}
\alias{get_artist_blogs}
\title{To get blogs about artist}
\usage{
get_artist_blogs(api_key, name = NA, id = NA, results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{id}{artist's id}

\item{results}{maximum size}
}
\value{
data frame giving blogs about artist
}
\description{
To get blogs about artist
}
\examples{
data=get_artist_blogs(api_key,name="coldplay",results=135)
}

