% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_images.R
\name{get_artist_images}
\alias{get_artist_images}
\title{To get artist's images}
\usage{
get_artist_images(api_key, name = NA, id = NA, start = NA, results = 15,
  license = "unknown")
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist name}

\item{id}{Echo Nest ID}

\item{start}{the desired index of the first result returned}

\item{results}{the number of results desired}

\item{license}{the desired licenses of the returned images}
}
\value{
data frame giving artist's images
}
\description{
To get artist's images
}
\examples{
\dontrun{
data=list_genres(api_key)
}
}

