% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_terms.R
\name{get_artist_terms}
\alias{get_artist_terms}
\title{To get artist's terms}
\usage{
get_artist_terms(api_key, name = NA, id = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{id}{artist's id}
}
\value{
data frame giving artist's terms
}
\description{
To get artist's terms
}
\examples{
\dontrun{
data=get_artist_terms(api_key,name="coldplay")
}
}

