% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_videos.R
\name{get_artist_videos}
\alias{get_artist_videos}
\title{To get a list of video documents found on the web related to an artist}
\usage{
get_artist_videos(api_key, name = NA, id = NA, start = NA, results = 15)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist name}

\item{id}{Echo Nest ID}

\item{start}{the desired index of the first result returned}

\item{results}{the number of results desired}
}
\value{
data frame giving artist's videos
}
\description{
To get a list of video documents found on the web related to an artist
}
\examples{
\dontrun{
data=get_artist_videos(api_key,name="coldplay")
}
}

