% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genre_info.R
\name{get_genre_info}
\alias{get_genre_info}
\title{To get basic information about a genre}
\usage{
get_genre_info(api_key, genre, description = T, urls = T)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{genre}{the genre name}

\item{description}{genre's description}

\item{urls}{genre's urls}
}
\value{
data frame giving basic info about a genre
}
\description{
To get basic information about a genre
}
\examples{
\dontrun{
data=get_genre_info(api_key,genre="post rock")
}
}

