% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_genre_artist.R
\name{get_top_genre_artists}
\alias{get_top_genre_artists}
\title{To Return the top artists for the given genre}
\usage{
get_top_genre_artists(api_key, genre)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{genre}{the genre name}
}
\value{
data frame top artist of the given genre
}
\description{
To Return the top artists for the given genre
}
\examples{
\dontrun{
data=get_top_genre_artists(api_key,genre="pop")
}
}

