% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_terms.R
\name{get_top_terms}
\alias{get_top_terms}
\title{To returns a list of the overall top terms}
\usage{
get_top_terms(api_key, results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{results}{the number of results desired}
}
\value{
data frame giving top terms
}
\description{
To returns a list of the overall top terms
}
\examples{
\dontrun{
data=get_top_terms(api_key)
}
}

