% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twitter_handle.R
\name{get_twitter_handle}
\alias{get_twitter_handle}
\title{To get the twitter handle for an artist}
\usage{
get_twitter_handle(api_key, name = NA, id = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist name}

\item{id}{Echo Nest ID}
}
\value{
data frame giving twitter handle
}
\description{
To get the twitter handle for an artist
}
\examples{
\dontrun{
data=get_twitter_handle(api_key,name="coldplay")
}
}

