% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_genres.R
\name{list_genres}
\alias{list_genres}
\title{To get genre's list}
\usage{
list_genres(api_key)
}
\arguments{
\item{api_key}{Echo Nest API key}
}
\value{
data frame giving genre's list
}
\description{
To get genre's list
}
\examples{
\dontrun{
data=list_genres(api_key)
}
}

