% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest2.R
\name{list_terms}
\alias{list_terms}
\title{To get a list of the best typed descriptive terms}
\usage{
list_terms(api_key, type = "style")
}
\arguments{
\item{api_key}{Echo Nest API key}
}
\value{
data frame giving best typed descriptive terms
}
\description{
To get a list of the best typed descriptive terms
}
\examples{
data=list_terms(api_key)
}

