% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_artist.R
\name{search_artist}
\alias{search_artist}
\title{To search artist by using name}
\usage{
search_artist(api_key, name = NA, style = NA, hotttnesss = T,
  description = NA, start = NA, results = 15, sort = NA, partner = NA,
  artist_location = NA, genre = NA, mood = NA, rank_type = "relevance",
  fuzzy_match = F, max_familiarity = NA, min_familiarity = NA,
  max_hotttnesss = NA, min_hotttnesss = NA, artist_start_year_before = NA,
  artist_start_year_after = NA, artist_end_year_before = NA,
  artist_end_year_after = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{style}{artist's style}

\item{hotttnesss}{artist's hotttnesss (Default is true)}

\item{description}{artist's description}

\item{start}{the desired index of the first result returned}

\item{results}{maximum size}

\item{sort}{to sort ascending or descending}

\item{partner}{partner catalog}

\item{artist_location}{artist location}

\item{genre}{genre name}

\item{mood}{mood like happy or sad}

\item{rank_type}{For search by description, style or mood indicates whether results should be ranked by query relevance or by artist familiarity}

\item{fuzzy_match}{if true, a fuzzy search is performed}

\item{max_familiarity}{maximum familiarity}

\item{min_familiarity}{minimum familiarity}

\item{max_hotttnesss}{maximum hotttnesss}

\item{min_hotttnesss}{minimum hotttnesss}

\item{artist_start_year_before}{Matches artists that have an earliest start year before the given value}

\item{artist_start_year_after}{Matches artists that have an earliest start year after the given value}

\item{artist_end_year_before}{Matches artists that have a latest end year before the given value}

\item{artist_end_year_after}{Matches artists that have a latest end year after the given value}
}
\value{
data frame giving artist's data
}
\description{
To search artist by using name
}
\examples{
\dontrun{
data=search_artist(api_key,"coldplay",sort="hotttnesss-desc",results=50)
}
}

