% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest.R
\name{search_artist_by_id}
\alias{search_artist_by_id}
\title{To search artist by using Echo Nest ID}
\usage{
search_artist_by_id(api_key, id)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{id}{Echo Nest ID}
}
\value{
data frame giving artist's data
}
\description{
To search artist by using Echo Nest ID
}
\examples{
data=search_artist_by_id(api_key,id="ARJ7KF01187B98D717" )
}

