% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest.R
\name{search_artist_by_name}
\alias{search_artist_by_name}
\title{To search artist by using name}
\usage{
search_artist_by_name(api_key, name = NA, style = NA, hotttnesss = T,
  description = NA, results = NA, sort = NA, partner = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{style}{artist's style}

\item{hotttnesss}{artist's hotttnesss (Default is true)}

\item{description}{artist's description}

\item{results}{maximum size}

\item{sort}{to sort ascending or descending}

\item{partner}{partner catalog}
}
\value{
data frame giving artist's data
}
\description{
To search artist by using name
}
\examples{
data=search_artist_by_name(api_key,"coldplay",sort="hotttnesss-desc",results=50)
}

