% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest.R
\name{search_songs}
\alias{search_songs}
\title{To search song}
\usage{
search_songs(api_key, name = NA, id = NA, hotttnesss = T, style = NA,
  title = NA, sort = NA, audio_summary = F, partner = NA,
  min_name = NA, max_name = NA, min_val = NA, max_val = NA,
  results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{id}{artist's id}

\item{hotttnesss}{song's hotttnesss}

\item{style}{artist's style}

\item{title}{song's title}

\item{sort}{to sort ascending or descending}

\item{audio_summary}{song's audio summary}

\item{partner}{partner catalog}

\item{min_name}{features' minimum value settings}

\item{max_name}{features' maximum value settings}

\item{min_val}{features' minimum value settings}

\item{max_val}{features' maximum value settings}

\item{results}{maximum size}
}
\value{
data frame giving artist's familiarity
}
\description{
To search song
}
\examples{
data=search_songs(api_key,style="pop",results=131)
}

