% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_artist.R
\name{similar_artists}
\alias{similar_artists}
\title{To search similar artists by using names or IDs}
\usage{
similar_artists(api_key, name = NA, id = NA, seed_catalog = NA,
  hotttnesss = T, start = 0, results = 15, max_familiarity = NA,
  min_familiarity = NA, max_hotttnesss = NA, min_hotttnesss = NA,
  artist_start_year_before = NA, artist_start_year_after = NA,
  artist_end_year_before = NA, artist_end_year_after = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artists' name (maximum upto 5 names)}

\item{id}{Echo Nest IDs (maximum upto 5 IDs)}

\item{seed_catalog}{seed catalog}

\item{hotttnesss}{artist's hotttnesss}

\item{start}{the desired index of the first result returned}

\item{results}{maximum size}

\item{max_familiarity}{maximum familiarity}

\item{min_familiarity}{minimum familiarity}

\item{max_hotttnesss}{maximum hotttnesss}

\item{min_hotttnesss}{minimum hotttnesss}

\item{artist_start_year_before}{Matches artists that have an earliest start year before the given value}

\item{artist_start_year_after}{Matches artists that have an earliest start year after the given value}

\item{artist_end_year_before}{Matches artists that have a latest end year before the given value}

\item{artist_end_year_after}{Matches artists that have a latest end year after the given value}

}
\value{
data frame giving similar artists' data
}
\description{
To search similar artists by using names or IDs
}
\examples{
\dontrun{
data=similar_artists(api_key,name=c("coldplay","adele","maroon 5"),results=35 )
}
}

