% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest.R
\name{similar_artists_by_ids}
\alias{similar_artists_by_ids}
\title{To search similar artists by using ids}
\usage{
similar_artists_by_ids(api_key, id = NA, hotttnesss = T, results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{id}{artists' id (maximum upto 5 ids)}

\item{hotttnesss}{artist's hotttnesss}

\item{results}{maximum size}
}
\value{
data frame giving similar artists' data
}
\description{
To search similar artists by using ids
}
\examples{
data=similar_artists_by_ids(api_key,id=c("ARJ7KF01187B98D717","AR7J9AP1187FB5BD64","ARF5M7Q1187FB501E8"),results=35 )
}

