% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rechonest.R
\name{similar_artists_by_names}
\alias{similar_artists_by_names}
\title{To search similar artists by using names}
\usage{
similar_artists_by_names(api_key, name = NA, hotttnesss = T, results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artists' name (maximum upto 5 names)}

\item{hotttnesss}{artist's hotttnesss}

\item{results}{maximum size}
}
\value{
data frame giving similar artists' data
}
\description{
To search similar artists by using names
}
\examples{
data=similar_artists_by_names(api_key,name=c("coldplay","adele","maroon 5"),results=35 )
}

