% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_playlist.R
\name{basic_playlist}
\alias{basic_playlist}
\title{To return basic playlist}
\usage{
basic_playlist(api_key, type = NA, artist_id = NA, artist = NA,
  song_id = NA, genre = NA, track_id = NA, results = 15, partner = NA,
  tracks = F, limited_interactivity = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{type}{the type of the playlist to be generated}

\item{artist_id}{artist id}

\item{artist}{artist name}

\item{song_id}{song ID}

\item{genre}{genre name}

\item{track_id}{track ID}

\item{results}{the number of results desired}

\item{partner}{partner catalog}

\item{tracks}{tracks info}

\item{limited_interactivity}{interactivity limitation}
}
\value{
data frame giving basic playlist
}
\description{
To return basic playlist
}
\examples{
\dontrun{
data=basic_playlist(api_key,type="artist-radio",artist=c("coldplay","adele"))
}
}

