% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_biographies.R
\name{get_artist_biographies}
\alias{get_artist_biographies}
\title{To get a list of artist biographies}
\usage{
get_artist_biographies(api_key, name = NA, id = NA, start = NA,
  results = 15, license = "unknown")
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist name}

\item{id}{Echo Nest ID}

\item{start}{the desired index of the first result returned}

\item{results}{the number of results desired}

\item{license}{the desired licenses of the returned images}
}
\value{
data frame giving artist's biographies
}
\description{
To get a list of artist biographies
}
\examples{
\dontrun{
data=get_artist_biographies(api_key,name="coldplay")
}
}

