% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_blogs.R
\name{get_artist_blogs}
\alias{get_artist_blogs}
\title{To get blogs about artist}
\usage{
get_artist_blogs(api_key, name = NA, start = NA, id = NA, results = 15,
  high_relevance = F)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{start}{the desired index of the first result returned}

\item{id}{artist's id}

\item{results}{maximum size}

\item{high_relevance}{if true only items that are highly relevant for this artist will be returned}
}
\value{
data frame giving blogs about artist
}
\description{
To get blogs about artist
}
\examples{
\dontrun{
data=get_artist_blogs(api_key,name="coldplay",results=35)
}
}

