% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_reviews.R
\name{get_artist_reviews}
\alias{get_artist_reviews}
\title{To get reviews about artist}
\usage{
get_artist_reviews(api_key, name = NA, id = NA, start = NA,
  results = 15)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{id}{artist's id}

\item{start}{the desired index of the first result returned}

\item{results}{maximum size}
}
\value{
data frame giving blogs about artist
}
\description{
To get reviews about artist
}
\examples{
\dontrun{
data=get_artist_reviews(api_key,name="coldplay",results=35)
}
}

