% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_songs.R
\name{get_artist_songs}
\alias{get_artist_songs}
\title{To get artist's songs}
\usage{
get_artist_songs(api_key, name = NA, id = NA, start = NA, results = 15)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{artist's name}

\item{id}{artist's id}

\item{start}{the desired index of the first result returned}

\item{results}{maximum size}
}
\value{
data frame giving artist's songs
}
\description{
To get artist's songs
}
\examples{
\dontrun{
data=get_artist_songs(api_key,name="coldplay")
}
}

