% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_hottt.R
\name{get_top_hottt}
\alias{get_top_hottt}
\title{To return a list of the top hottt artists}
\usage{
get_top_hottt(api_key, genre = NA, start = NA, results = 15)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{genre}{the set of genres of interest}

\item{start}{the desired index of the first result returned}

\item{results}{the number of results desired}
}
\value{
data frame giving top hottt artists
}
\description{
To return a list of the top hottt artists
}
\examples{
\dontrun{
data=get_top_hottt(api_key)
}
}

