% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_genre.R
\name{search_genre}
\alias{search_genre}
\title{To search for genres by name}
\usage{
search_genre(api_key, genre = NA, description = T, urls = T,
  results = 15)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{genre}{the genre name}

\item{description}{genre's description}

\item{urls}{genre's urls}

\item{results}{the number of results desired}
}
\value{
data frame giving searched genres
}
\description{
To search for genres by name
}
\examples{
\dontrun{
data=search_genre(api_key,genre="rock")\\
}
}

