% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_genres.R
\name{similar_genres}
\alias{similar_genres}
\title{To return similar genres to a given genre}
\usage{
similar_genres(api_key, genre = NA, description = T, urls = T,
  start = NA, results = 15)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{genre}{the genre name}

\item{description}{genre's description}

\item{urls}{genre's urls}

\item{start}{the desired index of the first result returned}

\item{results}{the number of results desired}
}
\value{
data frame giving similar genres
}
\description{
To return similar genres to a given genre
}
\examples{
\dontrun{
data=similar_genres(api_key,genre="rock")
}
}

