% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_artist_names.R
\name{suggest_artist_names}
\alias{suggest_artist_names}
\title{To suggest artists based upon partial names}
\usage{
suggest_artist_names(api_key, name, results = NA)
}
\arguments{
\item{api_key}{Echo Nest API key}

\item{name}{a partial artist name}

\item{results}{the number of results desired (maximum 15)}
}
\value{
data frame giving artist's names
}
\description{
To suggest artists based upon partial names
}
\examples{
\dontrun{
data=suggest_artist_names(api_key,"cold")
}
}

